Change matrices to use doubles instead of floats

+++ src/CCS.cpp
@@ -1,7 +1,7 @@
 #include "CCS.h"
 #include <cassert>
 
-CCSMatrix to_ccs(std::vector<std::vector<float>> mat)
+CCSMatrix to_ccs(std::vector<std::vector<double>> mat)
 {
     struct CCSMatrix out;
 
@@ -9,7 +9,7 @@ CCSMatrix to_ccs(std::vector<std::vector<float>> mat)
     out.num_cols = mat[0].size();
     out.non_zero = 0;
 
-    out.val = std::vector<float>();
+    out.val = std::vector<double>();
 
     for (unsigned int i = 0; i < out.num_cols; i++)
     {
@@ -32,10 +32,10 @@ CCSMatrix to_ccs(std::vector<std::vector<float>> mat)
     return out;
 }
 
-std::vector<std::vector<float>> from_ccs(CCSMatrix ccs)
+std::vector<std::vector<double>> from_ccs(CCSMatrix ccs)
 {
-    std::vector<std::vector<float>>
-        output(ccs.num_rows, std::vector<float>(ccs.num_cols, 0));
+    std::vector<std::vector<double>>
+        output(ccs.num_rows, std::vector<double>(ccs.num_cols, 0));
 
     for (unsigned int i = 0; i < ccs.col_ptr.size() - 1; i++)
     {
@@ -48,12 +48,12 @@ std::vector<std::vector<float>> from_ccs(CCSMatrix ccs)
     return output;
 }
 
-std::vector<float> ccs_vector_mult(CCSMatrix ccs, std::vector<float> vec)
+std::vector<double> ccs_vector_mult(CCSMatrix ccs, std::vector<double> vec)
 {
 
     assert(vec.size() == ccs.num_cols);
 
-    std::vector<float> output(ccs.num_rows, 0);
+    std::vector<double> output(ccs.num_rows, 0);
 
     for (unsigned int i = 0; i < ccs.col_ptr.size() - 1; i++)
     {
