+++ src/SampleMatrix.h
@@ -1,6 +1,6 @@
 #include <vector>
 
-std::vector<std::vector<float>> problem_1 = {
+std::vector<std::vector<double>> problem_1 = {
     {10., 0., 0., 0., -2., 0.},
     {3., 9., 0., 0., 0., 0.},
     {0., 0., 8., 7., 3., 0.},
